perf(check): Avoid recreating new App instances in unroll_app unnecessarily
Gives a ~65% speedup on the typecheck_prelude benchmark
